+++ bespokeasm/assembler/byte_code/assembled.py
@@ -3,6 +3,7 @@ import sys
 
 from bespokeasm.assembler.byte_code.parts import ByteCodePart
 from bespokeasm.assembler.byte_code.packed_bits import PackedBits
+from bespokeasm.assembler.label_scope import LabelScope
 
 class AssembledInstruction:
     def __init__(self, line_number: int, parts: list[ByteCodePart]):
@@ -29,10 +30,10 @@ class AssembledInstruction:
     def line_number(self):
         return self._line_number
 
-    def get_bytes(self, label_dict: dict[str, int]) -> bytearray:
+    def get_bytes(self, label_scope: LabelScope) -> bytearray:
         packed_bits = PackedBits()
         for p in self._parts:
-            value = p.get_value(self.line_number, label_dict)
+            value = p.get_value(self.line_number, label_scope)
             if  isinstance( value, str):
                 sys.exit(f'ERROR - assembled instruction "{self}" had a part {p} that did not resolve to an int, got: {value}')
             packed_bits.append_bits(