+++ test/test_assembler_model.py
@@ -2,6 +2,7 @@ import importlib.resources as pkg_resources
 import unittest
 import yaml
 
+from bespokeasm.assembler.label_scope import LabelScope, LabelScopeType
 import bespokeasm.assembler.model.operand as A
 import bespokeasm.assembler.model.operand_set as AS
 from bespokeasm.assembler.model import AssemblerModel
@@ -14,10 +15,15 @@ from test import config_files
 #
 
 class TestConfigObject(unittest.TestCase):
+    label_values = None
+
     @classmethod
     def setUpClass(cls):
         cls._register_argument_config_str = pkg_resources.read_text(config_files, 'register_argument_exmaple_config.yaml')
         cls._eater_sap1_config_str = pkg_resources.read_text(config_files, 'eater-sap1-isa.yaml')
+        cls.label_values = LabelScope(LabelScopeType.GLOBAL, None, 'global')
+        cls.label_values.set_label_value('label1', 2, 1)
+        cls.label_values.set_label_value('LABEL2', 0xF0, 2)
 
     def test_argument_set_construction(self):
         conf1 = yaml.safe_load(self._eater_sap1_config_str)
@@ -38,81 +44,76 @@ class TestConfigObject(unittest.TestCase):
         with pkg_resources.path(config_files, 'eater-sap1-isa.yaml') as fp:
             model1 = AssemblerModel(str(fp), 0)
 
-        label_dict = {
-            'label1': 2,
-            'LABEL2': 0xF0,
-        }
-
         pi1 = model1.parse_instruction(1212, 'LDA $f')
         self.assertEqual(pi1.byte_size, 1, 'assembled instruciton is 1 byte')
-        self.assertEqual(pi1.get_bytes(label_dict), bytearray([0x1F]), 'assembled instruction')
+        self.assertEqual(pi1.get_bytes(TestConfigObject.label_values), bytearray([0x1F]), 'assembled instruction')
 
         pi2 = model1.parse_instruction(1212, 'add label1+5')
         self.assertEqual(pi2.byte_size, 1, 'assembled instruciton is 1 byte')
-        self.assertEqual(pi2.get_bytes(label_dict), bytearray([0x27]), 'assembled instruction')
+        self.assertEqual(pi2.get_bytes(TestConfigObject.label_values), bytearray([0x27]), 'assembled instruction')
 
         pi3 = model1.parse_instruction(1212, 'out')
         self.assertEqual(pi3.byte_size, 1, 'assembled instruciton is 1 byte')
-        self.assertEqual(pi3.get_bytes(label_dict), bytearray([0xE0]), 'assembled instruction')
+        self.assertEqual(pi3.get_bytes(TestConfigObject.label_values), bytearray([0xE0]), 'assembled instruction')
 
         with pkg_resources.path(config_files, 'register_argument_exmaple_config.yaml') as fp:
             model2 = AssemblerModel(str(fp), 0)
 
         piA = model2.parse_instruction(1234, 'mov a, i')
         self.assertEqual(piA.byte_size, 1, 'assembled instruciton is 1 byte')
-        self.assertEqual(piA.get_bytes(label_dict), bytearray([0b01000010]), 'assembled instruction')
+        self.assertEqual(piA.get_bytes(TestConfigObject.label_values), bytearray([0b01000010]), 'assembled instruction')
 
         piB = model2.parse_instruction(1234, 'mov a,[$1120 + label1]')
         self.assertEqual(piB.byte_size, 3, 'assembled instruciton is 3 byte')
-        self.assertEqual(piB.get_bytes(label_dict), bytearray([0b01000110, 0x22, 0x11]), 'assembled instruction')
+        self.assertEqual(piB.get_bytes(TestConfigObject.label_values), bytearray([0b01000110, 0x22, 0x11]), 'assembled instruction')
 
         piC = model2.parse_instruction(888, 'add i')
         self.assertEqual(piC.byte_size, 1, 'assembled instruciton is 1 byte')
-        self.assertEqual(piC.get_bytes(label_dict), bytearray([0b10111010]), 'assembled instruction')
+        self.assertEqual(piC.get_bytes(TestConfigObject.label_values), bytearray([0b10111010]), 'assembled instruction')
 
         piD = model2.parse_instruction(1234, 'mov [$110D + (label1 + LABEL2)] , 0x88')
         self.assertEqual(piD.byte_size, 4, 'assembled instruciton is 4 byte')
-        self.assertEqual(piD.get_bytes(label_dict), bytearray([0b01110111,  0x88, 0xFF, 0x11]), 'arguments should be in reverse order')
+        self.assertEqual(piD.get_bytes(TestConfigObject.label_values), bytearray([0b01110111,  0x88, 0xFF, 0x11]), 'arguments should be in reverse order')
 
         piE = model2.parse_instruction(1234, 'mov [sp - label1] , 0x88')
         self.assertEqual(piE.byte_size, 3, 'assembled instruciton is 3 byte')
-        self.assertEqual(piE.get_bytes(label_dict), bytearray([0b01101111, 0x88, 0b11111110]), 'arguments should be in reverse order')
+        self.assertEqual(piE.get_bytes(TestConfigObject.label_values), bytearray([0b01101111, 0x88, 0b11111110]), 'arguments should be in reverse order')
 
         piF = model2.parse_instruction(1234, 'mov [sp+label1] , 0x88')
         self.assertEqual(piF.byte_size, 3, 'assembled instruciton is 3 byte')
-        self.assertEqual(piF.get_bytes(label_dict), bytearray([0b01101111, 0x88, 2]), 'arguments should be in reverse order')
+        self.assertEqual(piF.get_bytes(TestConfigObject.label_values), bytearray([0b01101111, 0x88, 2]), 'arguments should be in reverse order')
 
         piG = model2.parse_instruction(1234, 'mov [sp] , 0x88')
         self.assertEqual(piG.byte_size, 3, 'assembled instruciton is 3 byte')
-        self.assertEqual(piG.get_bytes(label_dict), bytearray([0b01101111, 0x88, 0]), 'arguments should be in reverse order')
+        self.assertEqual(piG.get_bytes(TestConfigObject.label_values), bytearray([0b01101111, 0x88, 0]), 'arguments should be in reverse order')
 
         piH = model2.parse_instruction(1234, 'mov [$8000], [label1]')
         self.assertEqual(piH.byte_size, 5, 'assembled instruciton is 5 byte')
-        self.assertEqual(piH.get_bytes(label_dict), bytearray([0b01110110, 2, 0, 0, 0x80]), 'arguments should be in reverse order')
+        self.assertEqual(piH.get_bytes(TestConfigObject.label_values), bytearray([0b01110110, 2, 0, 0, 0x80]), 'arguments should be in reverse order')
 
         piI = model2.parse_instruction(1234, 'mov [mar], [label1]')
         self.assertEqual(piI.byte_size, 3, 'assembled instruciton is 3 byte')
-        self.assertEqual(piI.get_bytes(label_dict), bytearray([0b01100110, 2, 0]), 'no offset should be emitted for [mar]')
+        self.assertEqual(piI.get_bytes(TestConfigObject.label_values), bytearray([0b01100110, 2, 0]), 'no offset should be emitted for [mar]')
 
         piJ = model2.parse_instruction(1234, 'swap [$8000], [label1]')
         self.assertEqual(piJ.byte_size, 5, 'assembled instruciton is 3 byte')
-        self.assertEqual(piJ.get_bytes(label_dict), bytearray([0b11110110, 0, 0x80, 2, 0]), 'arguments should NOT be in reverse order')
+        self.assertEqual(piJ.get_bytes(TestConfigObject.label_values), bytearray([0b11110110, 0, 0x80, 2, 0]), 'arguments should NOT be in reverse order')
 
         piK = model2.parse_instruction(1234, 'mov [sp+2], [sp+4]')
         self.assertEqual(piK.byte_size, 3, 'assembled instruciton is 3 byte')
-        self.assertEqual(piK.get_bytes(label_dict), bytearray([0b01101101, 4, 2]), 'arguments should be in reverse order')
+        self.assertEqual(piK.get_bytes(TestConfigObject.label_values), bytearray([0b01101101, 4, 2]), 'arguments should be in reverse order')
 
         piL = model2.parse_instruction(1234, 'pop i')
         self.assertEqual(piL.byte_size, 1, 'assembled instruciton is 1 byte')
-        self.assertEqual(piL.get_bytes(label_dict), bytearray([0b00001010]), 'pop to i')
+        self.assertEqual(piL.get_bytes(TestConfigObject.label_values), bytearray([0b00001010]), 'pop to i')
 
         piM = model2.parse_instruction(1234, 'pop')
         self.assertEqual(piM.byte_size, 1, 'assembled instruciton is 1 byte')
-        self.assertEqual(piM.get_bytes(label_dict), bytearray([0b00001111]), 'just pop')
+        self.assertEqual(piM.get_bytes(TestConfigObject.label_values), bytearray([0b00001111]), 'just pop')
 
         piN = model2.parse_instruction(1234, 'mov a, [sp+2]')
         self.assertEqual(piN.byte_size, 2, 'assembled instruciton is 2 byte')
-        self.assertEqual(piN.get_bytes(label_dict), bytearray([0b01000101, 2]), 'just move [sp+2] into a')
+        self.assertEqual(piN.get_bytes(TestConfigObject.label_values), bytearray([0b01000101, 2]), 'just move [sp+2] into a')
 
         with self.assertRaises(SystemExit, msg='should error on unallowed operand combinations'):
             model2.parse_instruction(1234, 'mov a, a')